fix: move top traders chain constants to Views/shared
Relocate SPOT_CHAINS, ALL_CHAINS, and PERP_CHAINS to the shared Views
module so SocialLeaderboard and Homepage can both import them without
violating ADR 0020 route isolation.
