+++ app/components/Views/shared/top-traders-constants.ts
